+++ omniva-woocommerce/includes/class-manifest.php
@@ -8,7 +8,16 @@ class OmnivaLt_Manifest
 
     wp_enqueue_script('moment', plugins_url('/js/moment.min.js', OmnivaLt_Core::$main_file_path), array(), null, true);
     wp_enqueue_script('bootstrap-datetimepicker', plugins_url('/js/datetimepicker/bootstrap-datetimepicker.min.js', OmnivaLt_Core::$main_file_path), array('jquery', 'moment'), null, true);
+    wp_enqueue_script('omniva_helper', plugins_url('/js/omniva_helper.js', OmnivaLt_Core::$main_file_path), array(), null, true);
     wp_enqueue_script('omniva_manifest', plugins_url('/js/omniva_manifest.js', OmnivaLt_Core::$main_file_path), array(), null, true);
+
+    wp_localize_script('omniva_manifest', 'omnivaglobals', array(
+      'cookie_checked_list' => 'omniva_checked',
+    ));
+
+    wp_localize_script('omniva_manifest', 'omnivatext', array(
+      'alert_select_orders' => __('Please select orders', 'omnivalt'),
+    ));
   }
 
   public static function register_menu_pages()