+++ omniva-woocommerce/js/omniva_helper.js
@@ -5,7 +5,7 @@ function omniva_setCookie(name,value,minutes) {
 		date.setTime(date.getTime() + (minutes*60*1000));
 		expires = "; expires=" + date.toUTCString();
 	}
-	document.cookie = name + "=" + (value || "")  + expires + "; path=/";
+	document.cookie = name + "=" + (value || "")  + expires + "; path=/;SameSite=Lax";
 }
 
 function omniva_getCookie(name) {