+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - added orders selection from multiple pages in Omniva Shipping page
 - labels are no longer stored in the plugin folder for a long time
 - added a ability to set custom shipping method name
+- updated Estonian translations
 
 ## [1.9.2.1]
 ### Fixed