+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added separation between delivery methods and delivery countries in settings page
 - added parameter "SameSite" for cookies when creating cookie via JavaSript
 - fixed mistake in COD code
+- separated plugin CSS animations
 
 ### Improved
 - added option to choose label design for delivery methods